This commit has major code changes and is the start of omapd-0.3.x branch.
Gone is the QtSOAP dependency!
Gone is the QtXML dependency!

QXmlStreamReader and QXmlStreamWriter are used for XML parsing and document
writing, respectively.  Custom classes have been created to marshall IF-MAP
requests and responses.  Metadata is stored as strings instead of QDomNodes.

Resolved is the issue with QDomDocument failing to report namespaces that are
not referenced in the XML document.  This caused a major issue with IF-MAP
operations that use a IF-MAP Filter, because namespace prefixes on the filter
have to be normalized against declared namespaces in the IF-MAP SOAP request.

All the client session information is passed around in a MapSessions singleton
object (the interface needs to be cleaned up though).  MapSessions holds lists
of clients with active polls, active ARC sessions, active SSRC sessions, and
client subscriptions.  It also tracks client configurations, though currently
only on-the-fly client configuration creation is supported.