Bug Fix for purgePublisher: Will release omapd-0.3.1.
Bug description: Fixed a regression where the publisher-id was not being added
to the Metadata object in publish requests in ClientParser::readMetadata(),
causing purgePublisher requests to fail matching any Metadata.  IF-MAP Filters
(for search, delete) still worked because the publisher-id was being written to
the Metadata xml string.

Slight tweak to certgen.sh to add a CN if none is supplied on command-line.