Bug Fix for search: Will release omapd-0.3.2.
Bug description: Fixed a regression where the metadata returned from a
QXmlQuery was not being trimmed (whitespace in front and end removed).  As a
result, metadata that did not match the query still had non-zero length, which
the search algorithm interpreted as having matched metadata on the link.  The
effect was a searchResult that contained identifiers and links that should not
be included in the searchResults.  Still, no metadata was leaked because the
QXmlQuery in effect still worked when the metadata was collected for the
searchResults.